# Conflicts:
#	pom.xml
#	src/main/java/com/cafe/crm/controllers/boss/StatisticController.java
#	src/main/java/com/cafe/crm/dto/ClientDetails.java
#	src/main/java/com/cafe/crm/services/impl/shift/ShiftServiceImpl.java
#	src/main/java/com/cafe/crm/services/impl/user/UserServiceImpl.java
#	src/main/java/com/cafe/crm/utils/RoundUpper.java

+++ src/main/java/com/cafe/crm/dto/ClientDetails.java
@@ -3,10 +3,10 @@ package com.cafe.crm.dto;
 
 public class ClientDetails {
 	private Double allDirtyPrice;
-	private Double otherPriceMenu;
-	private Double dirtyPriceMenu;
+	private Long otherPriceMenu;
+	private Long dirtyPriceMenu;
 
-	public ClientDetails(Double allDirtyPrice, Double otherPriceMenu, Double dirtyPriceMenu) {
+	public ClientDetails(Double allDirtyPrice, Long otherPriceMenu, Long dirtyPriceMenu) {
 		this.allDirtyPrice = allDirtyPrice;
 		this.otherPriceMenu = otherPriceMenu;
 		this.dirtyPriceMenu = dirtyPriceMenu;
@@ -20,19 +20,19 @@ public class ClientDetails {
 		this.allDirtyPrice = allDirtyPrice;
 	}
 
-	public Double getOtherPriceMenu() {
+	public Long getOtherPriceMenu() {
 		return otherPriceMenu;
 	}
 
-	public void setOtherPriceMenu(Double otherPriceMenu) {
+	public void setOtherPriceMenu(Long otherPriceMenu) {
 		this.otherPriceMenu = otherPriceMenu;
 	}
 
-	public Double getDirtyPriceMenu() {
+	public Long getDirtyPriceMenu() {
 		return dirtyPriceMenu;
 	}
 
-	public void setDirtyPriceMenu(Double dirtyPriceMenu) {
+	public void setDirtyPriceMenu(Long dirtyPriceMenu) {
 		this.dirtyPriceMenu = dirtyPriceMenu;
 	}
 }
